cli: share prompt_signal between update -i and init radio
Review feedback from @claude:

1. 'radio()' in 'bun init' got the ENABLE_PROCESSED_INPUT + drop(_restore)
   fixes but missed 'prompt_signal::install()', so external SIGINT during
   the init template picker still killed the process with the cursor
   hidden. Extract the module into 'src/runtime/cli/prompt_signal.rs'
   and install it in both call sites.

   The mouse-tracking-off bytes (\x1b[?1000l\x1b[?1006l) in the restore
   sequence are harmless for init (terminals ignore disable sequences
   for modes that were already off), so both sites share one string.

2. Intra-doc link '[PromptSignalGuard]' in the old inline module doc
   didn't resolve — the struct is named 'Guard'. Moot now that the
   module has its own file and top-level doc, but the prose was still
   wrong. Fixed as part of the move.

Verified:
- cargo check -p bun_runtime (linux) clean
- cargo check -p bun_runtime --target x86_64-pc-windows-msvc clean
- 'bun bd test' on update_interactive_install.test.ts + init.test.ts:
  15/15 pass.
