+++ src/runtime/cli/mod.rs
@@ -385,6 +385,8 @@ pub mod scan_command;
 pub mod unlink_command;
 #[path = "update_command.rs"]
 pub(crate) mod update_command;
+#[path = "prompt_signal.rs"]
+pub(crate) mod prompt_signal;
 #[path = "update_interactive_command.rs"]
 pub mod update_interactive_command;
 #[path = "why_command.rs"]
